BAC-5-3: add refactor of modal
