+++ src/shared/ui/VModal/model/types.ts
@@ -1,0 +1,4 @@
+export interface VModalAction {
+  show: CallableFunction;
+  hide: CallableFunction;
+}
